+++ apis/keda/v1alpha1/triggerauthentication_types.go
@@ -95,6 +95,9 @@ type TriggerAuthenticationSpec struct {
 
 	// +optional
 	AwsSecretManager *AwsSecretManager `json:"awsSecretManager,omitempty"`
+
+	// +optional
+	BoundServiceAccountToken []BoundServiceAccountToken `json:"boundServiceAccountToken,omitempty"`
 }
 
 // TriggerAuthenticationStatus defines the observed state of TriggerAuthentication
@@ -378,6 +381,11 @@ type AwsSecretManagerSecret struct {
 	VersionStage string `json:"versionStage,omitempty"`
 	// +optional
 	SecretKey string `json:"secretKey,omitempty"`
+}
+
+type BoundServiceAccountToken struct {
+	Parameter          string `json:"parameter"`
+	ServiceAccountName string `json:"serviceAccountName"`
 }
 
 func init() {